requirements.txt
python-dotenv==0.21.1
python_telegram_bot==20.0
requests==2.26.0
telegram==0.0.1
winotify==1.1.0
